requirements.txt
streamlit
azure-cognitiveservices-speech